+++ src/modules/ekf2/EKF/gps_checks.cpp
@@ -86,9 +86,9 @@ bool Ekf::collect_gps(const gps_message &gps)
 		const bool declination_was_valid = PX4_ISFINITE(_mag_declination_gps);
 
 		// set the magnetic field data returned by the geo library using the current GPS position
-		_mag_declination_gps = get_mag_declination_radians(lat, lon);
-		_mag_inclination_gps = get_mag_inclination_radians(lat, lon);
-		_mag_strength_gps = get_mag_strength_gauss(lat, lon);
+		_mag_declination_gps = get_mag_declination_radians(static_cast<float>(lat), static_cast<float>(lon));
+		_mag_inclination_gps = get_mag_inclination_radians(static_cast<float>(lat), static_cast<float>(lon));
+		_mag_strength_gps = get_mag_strength_gauss(static_cast<float>(lat), static_cast<float>(lon));
 
 		// request a reset of the yaw using the new declination
 		if ((_params.mag_fusion_type != MAG_FUSE_TYPE_NONE)
@@ -119,9 +119,9 @@ bool Ekf::collect_gps(const gps_message &gps)
 			const double lon = gps.lon * 1.0e-7;
 
 			// set the magnetic field data returned by the geo library using the current GPS position
-			_mag_declination_gps = get_mag_declination_radians(lat, lon);
-			_mag_inclination_gps = get_mag_inclination_radians(lat, lon);
-			_mag_strength_gps = get_mag_strength_gauss(lat, lon);
+			_mag_declination_gps = get_mag_declination_radians(static_cast<float>(lat), static_cast<float>(lon));
+			_mag_inclination_gps = get_mag_inclination_radians(static_cast<float>(lat), static_cast<float>(lon));
+			_mag_strength_gps = get_mag_strength_gauss(static_cast<float>(lat), static_cast<float>(lon));
 
 			// request mag yaw reset if there's a mag declination for the first time
 			if (_params.mag_fusion_type != MAG_FUSE_TYPE_NONE) {
